@@ -20,12 +20,7 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.google.common.collect.Lists;
-import com.google.common.util.concurrent.FutureCallback;
-import com.google.common.util.concurrent.Futures;
-import com.nostra13.universalimageloader.core.DisplayImageOptions;
-import com.nostra13.universalimageloader.core.ImageLoader;
-import com.nostra13.universalimageloader.core.assist.FailReason;
-import com.nostra13.universalimageloader.core.listener.ImageLoadingListener;
+
 
 import java.io.File;
 import java.util.HashSet;
@@ -34,12 +29,24 @@ import java.util.Set;
 
 import butterknife.ButterKnife;
 import butterknife.InjectView;
+import com.squareup.picasso.Callback;
+import com.squareup.picasso.Picasso;
+import retrofit.RestAdapter;
+import retrofit.converter.SimpleXMLConverter;
+import rx.Observable;
+import rx.android.schedulers.AndroidSchedulers;
+import rx.schedulers.Schedulers;
 import tn.opendata.tainan311.tainan1999.TainanReport1999;
-import tn.opendata.tainan311.tainan1999.rpc.QueryRequest;
+import tn.opendata.tainan311.tainan1999.api.QueryRequest;
+import tn.opendata.tainan311.tainan1999.api.Record;
+import tn.opendata.tainan311.tainan1999.api.Tainan1999Service;
+
 import tn.opendata.tainan311.tainan1999.util.TainanConstant;
-import tn.opendata.tainan311.tainan1999.vo.QueryResponse;
+
 import tn.opendata.tainan311.utils.LogUtils;
 import tn.opendata.tainan311.utils.PreferenceUtils;
+
+import static tn.opendata.tainan311.utils.EasyUtil.isNotEmpty;
 
 /**
  * Created by newman on 5/8/15.
@@ -48,20 +55,27 @@ public class MyActivity extends ListActivity {
     private static final String TAG = MyActivity.class.getSimpleName();
     // Object
     private QueryRequestArrayAdapter mQueryRequestArrayAdapter;
-    protected ImageLoader mImageLoader = ImageLoader.getInstance();
-    private DisplayImageOptions mOptions;
+
     // Value
-    private String mDataPath;
+    private RestAdapter restAdapter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         requestWindowFeature(Window.FEATURE_INDETERMINATE_PROGRESS);
+        setRawRequestIdsForTest();
+
+        restAdapter = new RestAdapter.Builder()
+                .setEndpoint("http://open1999.tainan.gov.tw:82")
+                .setConverter(new SimpleXMLConverter())
+                .setLogLevel(RestAdapter.LogLevel.FULL)
+                .build();
+
 
         initActionBar();
         initEmptyView();
         initView();
-        mDataPath = getFilesDir().toString()+"/pic/";
+
 
 //        setRawRequestIdsForTest();
         loadQueryRequest();
@@ -97,46 +111,49 @@ public class MyActivity extends ListActivity {
             LogUtils.d(TAG, "rids are ", rids);
             builder.setServiceRequestId(rids);
             LogUtils.d(TAG, "builder.build() is ", builder.build());
+            builder.setCityId(TainanConstant.CITY_ID);
 
-            Futures.addCallback(TainanReport1999.executeQuery(this, builder.build())
-                    , new FutureCallback<List<QueryResponse>>() {
-                @Override
-                public void onSuccess(final List<QueryResponse> result) {
-                    LogUtils.d(TAG, "callback onSuccess");
-                    runOnUiThread(new Runnable() {
-                        @Override
-                        public void run() {
-                            if (result != null && result.size() > 0) {
-                                LogUtils.d(TAG, "data count : ", result.size());
-                                if (mQueryRequestArrayAdapter == null) {
-                                    mQueryRequestArrayAdapter = new QueryRequestArrayAdapter(MyActivity.this, result);
-                                    setListAdapter(mQueryRequestArrayAdapter);
-                                    getListView().setOnItemClickListener(mQueryRequestArrayAdapter);
-                                } else {
-                                    mQueryRequestArrayAdapter.addAll(result);
-                                    mQueryRequestArrayAdapter.updateRequestList(result);
-                                }
-                            }
-                            showProgressOnActionBar(false);
-                        }
-                    });
-                }
+            Tainan1999Service service = restAdapter.create(Tainan1999Service.class);
+            service.queryReports(builder.build())
+                   .observeOn(AndroidSchedulers.mainThread())
+                   .subscribe(queryResponse -> {
+                       LogUtils.d(TAG, "callback onSuccess");
+                       if (queryResponse.getReturncode() == 0) { //success
+                           List<Record> records = queryResponse.getRecords();
+                           if (isNotEmpty(records)) {
+                               LogUtils.d(TAG, "data count : ", records.size());
+                               if (mQueryRequestArrayAdapter == null) {
+                                   mQueryRequestArrayAdapter = new QueryRequestArrayAdapter(MyActivity.this, records);
+                                   setListAdapter(mQueryRequestArrayAdapter);
+                                   getListView().setOnItemClickListener(mQueryRequestArrayAdapter);
+                               } else {
+                                   mQueryRequestArrayAdapter.addAll(records);
+                                   mQueryRequestArrayAdapter.updateRequestList(records);
+                               }
+                           }
 
-                @Override
-                public void onFailure(Throwable t) {
-                    // TODO
-                }
-            });
+
+                       } else {
+                           LogUtils.e(TAG, "error");
+                           //TODO: error handle??
+                       }
+                       showProgressOnActionBar(false);
+
+                   }, err -> {
+                       showProgressOnActionBar(false);
+                       LogUtils.e(TAG, err.getMessage());
+                   });
+
         } else {
             LogUtils.d(TAG, "requestIds is null or empty");
         }
     }
 
-    public class QueryRequestArrayAdapter extends ArrayAdapter<QueryResponse> implements AdapterView.OnItemClickListener {
+    public class QueryRequestArrayAdapter extends ArrayAdapter<Record> implements AdapterView.OnItemClickListener {
         private final LayoutInflater mInflater;
         private final int mResource;
-        private final List<QueryResponse> mRequestList = Lists.newArrayList();
-        private QueryRequestArrayAdapter(Context context, List<QueryResponse> objects) {
+        private final List<Record> mRequestList = Lists.newArrayList();
+        private QueryRequestArrayAdapter(Context context, List<Record> objects) {
             super(context, R.layout.list_item_request, objects);
             mInflater = (LayoutInflater) context.getSystemService(Service.LAYOUT_INFLATER_SERVICE);
             mResource = R.layout.list_item_request;
@@ -150,33 +167,37 @@ public class MyActivity extends ListActivity {
                 ViewHolder holder = new ViewHolder(convertView);
                 convertView.setTag(holder);
             }
-            QueryResponse r = getItem(position);
+            Record r = getItem(position);
             final ViewHolder holder = (ViewHolder) convertView.getTag();
             // image
-            File file = new File(mDataPath+r.getService_request_id()+".jpg");
-            if (file.exists()) {
-                mImageLoader.loadImage(Uri.fromFile(file).toString(), mOptions, new ImageLoadingListener() {
-                    @Override
-                    public void onLoadingStarted(String imageUri, View view) {
-                    }
+            holder.cover.setVisibility(View.GONE);
+            Observable.from(r.getPictures())
+                      .observeOn(Schedulers.io())
+                      .filter(pic -> !TextUtils.isEmpty(pic.getFile()) || pic.getFilePath() != null)
+                      .doOnNext(pic -> pic.doPrepareImage(getContext()))
+                    .first() //We only need 1 pic
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .subscribe(pic -> {
+                                Picasso.with(getContext())
+                                       .load(new File(pic.getFilePath()))
+                                       .fit()
+                                       .centerCrop()
+                                       .into(holder.cover, new Callback() {
+                                           @Override
+                                           public void onSuccess() {
+                                               holder.cover.setVisibility(View.VISIBLE);
+                                           }
 
-                    @Override
-                    public void onLoadingFailed(String imageUri, View view, FailReason failReason) {
-                    }
+                                           @Override
+                                           public void onError() {
+                                               LogUtils.w(TAG, "onError");
+                                           }
+                                       });
+                            }, err -> LogUtils.e(TAG, err.getMessage())
 
-                    @Override
-                    public void onLoadingComplete(String imageUri, View view, Bitmap loadedImage) {
-                        holder.cover.setVisibility(View.VISIBLE);
-                        holder.cover.setImageBitmap(loadedImage);
-                    }
+                    );
 
-                    @Override
-                    public void onLoadingCancelled(String imageUri, View view) {
-                    }
-                });
-            } else {
-                holder.cover.setVisibility(View.GONE);
-            }
+
             // service name
             holder.service_name.setText(r.getService_name());
             // subproject
@@ -226,7 +247,7 @@ public class MyActivity extends ListActivity {
             startActivity(i);
         }
 
-        public void updateRequestList(List<QueryResponse> list) {
+        public void updateRequestList(List<Record> list) {
             mRequestList.addAll(list);
             mQueryRequestArrayAdapter.notifyDataSetChanged();
         }
