@@ -5,30 +5,27 @@ import android.app.ListActivity;
 import android.app.Service;
 import android.content.Context;
 import android.content.Intent;
-import android.graphics.Bitmap;
-import android.net.Uri;
 import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.AbsListView;
-import android.widget.AdapterView;
-import android.widget.ArrayAdapter;
-import android.widget.ImageView;
-import android.widget.LinearLayout;
-import android.widget.TextView;
-
+import android.text.TextUtils;
+import android.view.*;
+import android.widget.*;
+import butterknife.ButterKnife;
+import butterknife.InjectView;
+import butterknife.OnClick;
 import com.getbase.floatingactionbutton.AddFloatingActionButton;
 import com.google.common.collect.Lists;
-import com.google.common.util.concurrent.FutureCallback;
-import com.google.common.util.concurrent.Futures;
-import com.nostra13.universalimageloader.core.DisplayImageOptions;
-import com.nostra13.universalimageloader.core.ImageLoader;
-import com.nostra13.universalimageloader.core.ImageLoaderConfiguration;
-import com.nostra13.universalimageloader.core.assist.FailReason;
-import com.nostra13.universalimageloader.core.listener.ImageLoadingListener;
+import com.squareup.picasso.Callback;
+import com.squareup.picasso.Picasso;
+import retrofit.RestAdapter;
+import retrofit.converter.SimpleXMLConverter;
+import rx.Observable;
+import rx.android.schedulers.AndroidSchedulers;
+import rx.schedulers.Schedulers;
+import tn.opendata.tainan311.tainan1999.api.QueryRequest;
+import tn.opendata.tainan311.tainan1999.api.Record;
+import tn.opendata.tainan311.tainan1999.api.Tainan1999Service;
+import tn.opendata.tainan311.tainan1999.util.TainanConstant;
+import tn.opendata.tainan311.utils.LogUtils;
 
 import java.io.File;
 import java.text.ParseException;
@@ -37,42 +34,60 @@ import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
 
-import butterknife.ButterKnife;
-import butterknife.InjectView;
-import butterknife.OnClick;
-import tn.opendata.tainan311.tainan1999.TainanReport1999;
-import tn.opendata.tainan311.tainan1999.rpc.QueryRequest;
-import tn.opendata.tainan311.tainan1999.util.TainanConstant;
-import tn.opendata.tainan311.tainan1999.vo.QueryResponse;
-import tn.opendata.tainan311.utils.LogUtils;
+import static tn.opendata.tainan311.tainan1999.api.QueryRequest.Builder;
+
 
 /**
  * Created by newman on 5/5/15.
  */
 public class TainanRequestListActivity extends ListActivity {
-    @InjectView(R.id.normal_plus) AddFloatingActionButton addRequestButton;
     private static final String TAG = TainanRequestListActivity.class.getSimpleName();
+    @InjectView(R.id.normal_plus)
+    AddFloatingActionButton addRequestButton;
     // View
     private LinearLayout mLoadingMoreItem;
     // Object
     private QueryRequestArrayAdapter mQueryRequestArrayAdapter;
-    protected ImageLoader mImageLoader = ImageLoader.getInstance();
-    private DisplayImageOptions mOptions;
+
+
     private SimpleDateFormat mSimpleDateFormat;
     // Value
     private boolean mLoadingMore;
-    private String mDataPath;
+
     // Constant
+    private RestAdapter restAdapter;
+    private Calendar cal = Calendar.getInstance();
+    private AbsListView.OnScrollListener mOnScrollListener = new AbsListView.OnScrollListener() {
+        @Override
+        public void onScrollStateChanged(AbsListView absListView, int i) {
+        }
+
+        @Override
+        public void onScroll(AbsListView view, int firstVisibleItem, int visibleItemCount, int totalItemCount) {
+            int lastInScreen = firstVisibleItem + visibleItemCount;
+            //is the bottom item visible & not loading more already ? Load more !
+            if ((lastInScreen == totalItemCount) && !(mLoadingMore)) {
+                //TODO send a another request
+                loadQueryRequest(false);
+            }
+        }
+    };
 
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_tainan_query_list);
         ButterKnife.inject(this);
         mSimpleDateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss", getResources().getConfiguration().locale);
-        mDataPath = getFilesDir().toString()+"/pic/";
+
+
+        restAdapter = new RestAdapter.Builder()
+                .setEndpoint("http://open1999.tainan.gov.tw:82")
+                .setConverter(new SimpleXMLConverter())
+                .setLogLevel(RestAdapter.LogLevel.FULL)
+                .build();
+
 
         mLoadingMore = false;
-        initImageLoader();
         initActionBar();
         initViews();
         loadQueryRequest(true);
@@ -86,7 +101,8 @@ public class TainanRequestListActivity extends ListActivity {
 
     private void initViews() {
         getListView().setDivider(null);
-        mLoadingMoreItem = (LinearLayout) ((LayoutInflater) getSystemService(LAYOUT_INFLATER_SERVICE)).inflate(R.layout.list_item_loading_more, null);
+        mLoadingMoreItem = (LinearLayout) ((LayoutInflater) getSystemService(LAYOUT_INFLATER_SERVICE))
+                .inflate(R.layout.list_item_loading_more, null);
         getListView().addFooterView(mLoadingMoreItem);
     }
 
@@ -95,13 +111,10 @@ public class TainanRequestListActivity extends ListActivity {
         startActivity(new Intent(TainanRequestListActivity.this, ReportActivity.class));
     }
 
-    private Calendar cal = Calendar.getInstance();
     private void loadQueryRequest(boolean firstTimeQuery) {
         LogUtils.d(TAG, "loadQueryRequest firstTimeQuery:", firstTimeQuery);
         addLoadingMoreListItem();
-
-        QueryRequest.Builder builder = QueryRequest.Builder.create();
-        builder.setEndTime(mSimpleDateFormat.format(cal.getTime()));
+        Date endDate = cal.getTime();
         if (firstTimeQuery) {
             cal.set(Calendar.HOUR, 0);
             cal.set(Calendar.MINUTE, 0);
@@ -109,50 +122,51 @@ public class TainanRequestListActivity extends ListActivity {
         } else {
             cal.add(Calendar.DAY_OF_YEAR, -1);
         }
-        builder.setStartTime(mSimpleDateFormat.format(cal.getTime()));
-        LogUtils.d(TAG, "builder.build() is ", builder.build());
+
+        QueryRequest request = Builder.create()
+                                      .setCityId(TainanConstant.CITY_ID)
+                                      .setEndDate(mSimpleDateFormat.format(endDate))
+                                      .setStartDate(mSimpleDateFormat.format(cal.getTime()))
+                                      .build();
+
+        // LogUtils.d(TAG, "builder.build() is ", builder());
 
         mLoadingMore = true;
-        Futures.addCallback(TainanReport1999.executeQuery(this, builder.build())
-                , new FutureCallback<List<QueryResponse>>() {
-            @Override
-            public void onSuccess(final List<QueryResponse> result) {
-                LogUtils.d(TAG, "callback onSuccess");
-                mLoadingMore = false;
-                runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        removeLoadingMoreListItem();
-                        if (result != null && result.size() > 0) {
-                            LogUtils.d(TAG, "data count : ", result.size());
-                            if (mQueryRequestArrayAdapter == null) {
-                                mQueryRequestArrayAdapter = new QueryRequestArrayAdapter(TainanRequestListActivity.this, result);
-                                setListAdapter(mQueryRequestArrayAdapter);
-                                getListView().setOnItemClickListener(mQueryRequestArrayAdapter);
-                                getListView().setOnScrollListener(mOnScrollListener);
-                            } else {
-                                mQueryRequestArrayAdapter.addAll(result);
-                                mQueryRequestArrayAdapter.updateRequestList(result);
-                            }
-                        }
-                    }
-                });
-            }
 
-            @Override
-            public void onFailure(Throwable t) {
-                // TODO
-            }
-        });
-    }
+        Tainan1999Service service = restAdapter.create(Tainan1999Service.class);
+        service.queryReports(request)
+               .observeOn(AndroidSchedulers.mainThread())
+               .subscribe(queryResponse -> {
+                   LogUtils.d(TAG, "callback onSuccess");
+                   mLoadingMore = false;
+                   removeLoadingMoreListItem();
+                   if (queryResponse.getReturncode() == 0) { //success
+                       List<Record> records = queryResponse.getRecords();
+                       if (mQueryRequestArrayAdapter == null) {
+                           mQueryRequestArrayAdapter = new QueryRequestArrayAdapter(TainanRequestListActivity.this, records);
+                           setListAdapter(mQueryRequestArrayAdapter);
+                           getListView().setOnItemClickListener(mQueryRequestArrayAdapter);
+                           getListView().setOnScrollListener(mOnScrollListener);
+                       } else {
+                           mQueryRequestArrayAdapter.addAll(records);
+                           mQueryRequestArrayAdapter.updateRequestList(records);
+                       }
 
-    private void initImageLoader() {
-        mImageLoader.init(ImageLoaderConfiguration.createDefault(this));
-        mOptions = new DisplayImageOptions.Builder()
-                .cacheInMemory(true)
-                .cacheOnDisk(true)
-                .bitmapConfig(Bitmap.Config.RGB_565)
-                .build();
+
+                       //FIXME: remove it...
+                       if (records != null && records.size() > 0) {
+                           LogUtils.d(TAG, "data count : ", records.size());
+                       }
+
+                   } else {
+                       LogUtils.e(TAG, "error");
+                       //TODO: error handle??
+                   }
+
+
+               }, err -> {
+                   LogUtils.e(TAG, err.getMessage());
+               });
     }
 
     private void addLoadingMoreListItem() {
@@ -161,104 +175,6 @@ public class TainanRequestListActivity extends ListActivity {
 
     private void removeLoadingMoreListItem() {
         mLoadingMoreItem.setVisibility(View.GONE);
-    }
-
-    public class QueryRequestArrayAdapter extends ArrayAdapter<QueryResponse> implements AdapterView.OnItemClickListener {
-        private final LayoutInflater mInflater;
-        private final int mResource;
-        private final List<QueryResponse> mRequestList = Lists.newArrayList();
-        private QueryRequestArrayAdapter(Context context, List<QueryResponse> objects) {
-            super(context, R.layout.list_item_request, objects);
-            mInflater = (LayoutInflater) context.getSystemService(Service.LAYOUT_INFLATER_SERVICE);
-            mResource = R.layout.list_item_request;
-            mRequestList.addAll(objects);
-        }
-
-        @Override
-        public View getView(int position, View convertView, ViewGroup parent) {
-            if (convertView == null) {
-                convertView = mInflater.inflate(mResource, parent, false);
-                ViewHolder holder = new ViewHolder(convertView);
-                convertView.setTag(holder);
-            }
-            QueryResponse r = getItem(position);
-            final ViewHolder holder = (ViewHolder) convertView.getTag();
-            // image
-            File file = new File(mDataPath+r.getService_request_id()+".jpg");
-            if (file.exists()) {
-                mImageLoader.loadImage(Uri.fromFile(file).toString(), mOptions, new ImageLoadingListener() {
-                    @Override
-                    public void onLoadingStarted(String imageUri, View view) {
-                    }
-
-                    @Override
-                    public void onLoadingFailed(String imageUri, View view, FailReason failReason) {
-                    }
-
-                    @Override
-                    public void onLoadingComplete(String imageUri, View view, Bitmap loadedImage) {
-                        holder.cover.setVisibility(View.VISIBLE);
-                        holder.cover.setImageBitmap(loadedImage);
-                    }
-
-                    @Override
-                    public void onLoadingCancelled(String imageUri, View view) {
-                    }
-                });
-            } else {
-                holder.cover.setVisibility(View.GONE);
-            }
-            // service name
-            holder.service_name.setText(r.getService_name());
-            // subproject
-            holder.subproject.setText(r.getSubproject());
-            // area
-            holder.area.setText(r.getArea());
-            // status
-            if (TainanConstant.STATUS_FINISH.equals(r.getStatus())) {
-                holder.status.setText(R.string.status_finished);
-                holder.status.setBackgroundColor(getResources().getColor(android.R.color.holo_green_light));
-                holder.status.setVisibility(View.VISIBLE);
-            } else if (TainanConstant.STATUS_IN_PROCESS.equals(r.getStatus())) {
-                holder.status.setText(R.string.status_inprogress);
-                holder.status.setBackgroundColor(getResources().getColor(android.R.color.holo_orange_dark));
-                holder.status.setVisibility(View.VISIBLE);
-            } else if (TainanConstant.STATUS_NOT_TAKEN.equals(r.getStatus())) {
-                holder.status.setText(R.string.status_not_taken);
-                holder.status.setBackgroundColor(getResources().getColor(android.R.color.holo_red_light));
-                holder.status.setVisibility(View.VISIBLE);
-            } else {
-                holder.status.setVisibility(View.GONE);
-            }
-            // requested date and time
-            holder.datetime.setText(r.getRequested_datetime());
-            return convertView;
-        }
-
-        public class ViewHolder {
-            public ViewHolder(View view) {
-                ButterKnife.inject(this, view);
-            }
-
-            @InjectView(R.id.img) ImageView cover;
-            @InjectView(R.id.subproject) TextView subproject;
-            @InjectView(R.id.service_name) TextView service_name;
-            @InjectView(R.id.datetime) TextView datetime;
-            @InjectView(R.id.status) TextView status;
-            @InjectView(R.id.area) TextView area;
-        }
-
-        @Override
-        public void onItemClick(AdapterView<?> adapterView, View view, int position, long id) {
-            Intent i = new Intent(TainanRequestListActivity.this, DetailActivity.class);
-            i.putExtra(DetailActivity.EXTRA_KEY_REQUEST, getItem(position));
-            startActivity(i);
-        }
-
-        public void updateRequestList(List<QueryResponse> list) {
-            mRequestList.addAll(list);
-            mQueryRequestArrayAdapter.notifyDataSetChanged();
-        }
     }
 
     @Override
@@ -293,27 +209,116 @@ public class TainanRequestListActivity extends ListActivity {
         return true;
     }
 
-    private AbsListView.OnScrollListener mOnScrollListener = new AbsListView.OnScrollListener() {
-        @Override
-        public void onScrollStateChanged(AbsListView absListView, int i) {
+    public class QueryRequestArrayAdapter extends ArrayAdapter<Record> implements AdapterView.OnItemClickListener {
+        private final LayoutInflater mInflater;
+        private final int mResource;
+        private final List<Record> mRequestList = Lists.newArrayList();
+
+        private QueryRequestArrayAdapter(Context context, List<Record> objects) {
+            super(context, R.layout.list_item_request, objects);
+            mInflater = (LayoutInflater) context.getSystemService(Service.LAYOUT_INFLATER_SERVICE);
+            mResource = R.layout.list_item_request;
+            mRequestList.addAll(objects);
         }
 
         @Override
-        public void onScroll(AbsListView view, int firstVisibleItem, int visibleItemCount, int totalItemCount) {
-            int lastInScreen = firstVisibleItem + visibleItemCount;
-            //is the bottom item visible & not loading more already ? Load more !
-            if((lastInScreen == totalItemCount) && !(mLoadingMore)){
-                QueryResponse r = mQueryRequestArrayAdapter.getItem(mQueryRequestArrayAdapter.getCount()-1);
-                String time = r.getRequested_datetime();
-                try {
-                    Date d = mSimpleDateFormat.parse(time);
-                    cal.setTimeInMillis(d.getTime());
-                    LogUtils.d(TAG, "time is ", time, "  cal is ", cal.toString());
-                    loadQueryRequest(false);
-                } catch (ParseException e) {
-                    LogUtils.w(TAG, e.getMessage(), e);
-                }
+        public View getView(int position, View convertView, ViewGroup parent) {
+            if (convertView == null) {
+                convertView = mInflater.inflate(mResource, parent, false);
+                ViewHolder holder = new ViewHolder(convertView);
+                convertView.setTag(holder);
             }
+
+            Record r = getItem(position);
+            final ViewHolder holder = (ViewHolder) convertView.getTag();
+            // image
+            holder.cover.setVisibility(View.INVISIBLE);
+            Observable.from(r.getPictures())
+                      .observeOn(Schedulers.io())
+                      .filter(pic -> !TextUtils.isEmpty(pic.getFile()) || pic.getFilePath() != null)
+                      .doOnNext(pic -> pic.doPrepareImage(getContext()))
+                      .first() //We only need 1 pic
+                      .observeOn(AndroidSchedulers.mainThread())
+                      .subscribe(pic -> {
+                                Picasso.with(getContext())
+                                       .load(new File(pic.getFilePath()))
+                                       .fit()
+                                       .centerCrop()
+                                       .into(holder.cover, new Callback() {
+                                           @Override
+                                           public void onSuccess() {
+                                               holder.cover.setVisibility(View.VISIBLE);
+                                           }
+
+                                           @Override
+                                           public void onError() {
+                                               LogUtils.w(TAG, "onError");
+                                           }
+                                       });
+                            }, err -> LogUtils.e(TAG, err.getMessage())
+
+                    );
+
+            // service name
+            holder.service_name.setText(r.getService_name());
+            // subproject
+            holder.subproject.setText(r.getSubproject());
+            // area
+            holder.area.setText(r.getArea());
+            // status
+            if (TainanConstant.STATUS_FINISH.equals(r.getStatus())) {
+                holder.status.setText(R.string.status_finished);
+                holder.status.setBackgroundColor(getResources().getColor(android.R.color.holo_green_light));
+                holder.status.setVisibility(View.VISIBLE);
+            } else if (TainanConstant.STATUS_IN_PROCESS.equals(r.getStatus())) {
+                holder.status.setText(R.string.status_inprogress);
+                holder.status.setBackgroundColor(getResources().getColor(android.R.color.holo_orange_dark));
+                holder.status.setVisibility(View.VISIBLE);
+            } else if (TainanConstant.STATUS_NOT_TAKEN.equals(r.getStatus())) {
+                holder.status.setText(R.string.status_not_taken);
+                holder.status.setBackgroundColor(getResources().getColor(android.R.color.holo_red_light));
+                holder.status.setVisibility(View.VISIBLE);
+            } else {
+                holder.status.setVisibility(View.GONE);
+            }
+            // requested date and time
+            holder.datetime.setText(r.getRequested_datetime());
+            return convertView;
         }
-    };
+
+        @Override
+        public void onItemClick(AdapterView<?> adapterView, View view, int position, long id) {
+            Intent i = new Intent(TainanRequestListActivity.this, DetailActivity.class);
+            i.putExtra(DetailActivity.EXTRA_KEY_REQUEST, getItem(position));
+            startActivity(i);
+        }
+
+        public void updateRequestList(List<Record> list) {
+            mRequestList.addAll(list);
+            mQueryRequestArrayAdapter.notifyDataSetChanged();
+        }
+
+    }
+
+
+    public class ViewHolder {
+        @InjectView(R.id.img)
+        ImageView cover;
+        @InjectView(R.id.subproject)
+        TextView subproject;
+        @InjectView(R.id.service_name)
+        TextView service_name;
+        @InjectView(R.id.datetime)
+        TextView datetime;
+        @InjectView(R.id.status)
+        TextView status;
+        @InjectView(R.id.area)
+        TextView area;
+
+
+        public ViewHolder(View view) {
+            ButterKnife.inject(this, view);
+
+        }
+    }
 }
